Remove unneeded casts.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/bcel/trunk@1702134 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/bcel6/classfile/StackMapType.java
@@ -47,7 +47,7 @@ public final class StackMapType implements Cloneable {
     StackMapType(DataInput file, ConstantPool constant_pool) throws IOException {
         this(file.readByte(), -1, constant_pool);
         if (hasIndex()) {
-            this.index = (int) file.readShort();
+            this.index = file.readShort();
         }
         this.constant_pool = constant_pool;
     }
